更新代码

+++ EveryStamp/Models/Remote/Request/Requests/UserRequest/UserChangePwdRequest.swift
@@ -16,7 +16,7 @@ struct UserChangePwdRequest: BaseRequest {
     
     var paramater: [String : Any]?
     
-    init(accessToken: String, userId: Double,iniPwd: String, pwd: String) {
+    init(accessToken: String, userId: Int, iniPwd: String, pwd: String) {
         paramater = ["access_token": accessToken , "user_id": userId, "ini_pwd": iniPwd,"pwd": pwd]
     }
 }
