+++ EveryStamp/Models/Remote/Request/Requests/UserRequest/UserRegisterRequest.swift
@@ -18,7 +18,7 @@ struct UserRegisterRequest: BaseRequest {
     
     var paramater: [String : Any]?
     
-    init(time: String, token: String, userName: String, pwd: String, code: Double, from: String) {
+    init(time: String, token: String, userName: String, pwd: String, code: Int, from: String) {
         paramater = ["time": time , "token": token, "user_name": userName,"pwd": pwd,"code": code, "from": from]
     }
